backend/transcription_service/requirements.txt
# Base dependencies
fastapi>=0.111.1
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
python-multipart>=0.0.9 # For file uploads

# WhisperX and its core dependencies
# Using git+https to potentially get latest fixes/features
whisperx @ git+https://github.com/m-bain/whisperX.git

# Faster-whisper is usually a dependency of whisperx, but explicitly listing
# faster-whisper>=0.10.0 # Consider pinning if needed

# Pyannote for diarization
pyannote.audio>=3.1.1 # Or pin to a version known compatible with whisperx

# Torch dependencies (ensure compatibility with Docker image CUDA version)
# Using cu126 index as requested (suitable for CUDA >= 12.6)
torch>=2.1.0 --index-url https://download.pytorch.org/whl/cu126
torchaudio>=2.1.0 --index-url https://download.pytorch.org/whl/cu126

# If using CPU only, you might use:
# torch>=2.1.0
# torchaudio>=2.1.0

# Setuptools-rust is often needed for building whisper dependencies
setuptools-rust>=1.1.2
onnxruntime-gpu # For GPU execution with some alignment/diarization backends 